AWS S3 adapter using the bare aws-sdk gem
An alternate adapter to the S3Adapter for those using the plain old
bare 'aws-sdk' gem without 'fog'. This class modelled on that adapter
but also sets up the S3 object to return the following additional HTTP
headers:

* Cache-Control: private, max-age=0, no-cache
* Content-Type: application/x-gzip

for completeness.

No integration test created, as noted in the s3_adapter_spec.rb, it
would be of limited use.

IAM authentication is not yet supported.