AwsSdkAdapter default using credential resolver
In the current behavior,
you can not use instance profiles and task role
in the instance that creates the sitemap.

aws-sdk-ruby v2 is auto detected using credentials.

According to the behavior of aws-sdk-ruby v2
unless you specify anything,
explicitly specify that you use the specified credentials.

Backwards compatibility is preserved because
environment variables are used in the
default behavior of aws-sdk-ruby v2.